feat: `splice`
