+++ src/index.ts
@@ -77,7 +77,7 @@ export * from './randomString';
 export * from './range';
 export * from './reduce';
 export * from './reject';
-export * from './remove';
+export * from './splice';
 export * from './reverse';
 export * from './sample';
 export * from './set';
